refactor(generation): split lidBuilder into focused modules
Decompose the 841-LOC lidBuilder.ts into 5 sibling modules organized by
geometry phase. The orchestrator becomes a 96-LOC `buildLid` that wires
the phases together with a single `withScope`.

Extracted modules:

- lidInputs.ts (125) — `LidInputs` interface + `resolveLidInputs` (the
  conversion from BinParams to the lid-local geometry tuple).
- lidProfile.ts (107) — `buildOutlineDrawing` (rounded-rect or polygon
  perimeter at a given inset), `buildMatingShell` (inverted-lip wall via
  outer/inner lofts + cut), `buildLidFloor` (flat plate at the top).
- lidClickRail.ts (296) — `chamferApexXForCavityWall` +
  `clickShape2D` + `buildClickRailBar` + `RailPlacement` interface +
  `railPlacementsForPolygon` + `railPlacementsForRectangle` +
  `addClickRails`. The full per-wall snap-rail pipeline.
- lidStackGrid.ts (122) — `buildLidStackPocketCutter`,
  `buildStackGrid`, `isCellFilled` (the predicate also used by the
  magnet pass for polygon bins, hence the export).
- lidMagnets.ts (72) — `cutMagnetHoles` (blind cylindrical pocket
  pattern with the LID_MAGNET_CEILING + topThickness floor cap).

lidBuilder.ts: 841 LOC -> 96 LOC. The two test-only helpers
(`resolveLidInputs`, `chamferApexXForCavityWall`) are re-exported so
lidBuilder.test.ts compiles unchanged. `buildLid` (the only consumer
in production code) signature is byte-identical.

2 lid test files / 34 tests pass (11.7s targeted run).
